refactor: 채플 수강 학점 공통 교양 학점 반영 (#202)
* refactor: 채플 학점 1회당 0.5점 반영

* refactor: 채플 수강에 따른 수강 학점 공통교양에 수강 학점에 반영

* refactor: Response type 수정

+++ src/main/java/com/plzgraduate/myongjigraduatebe/graduation/application/port/in/response/BasicInfoResponse.java
@@ -13,10 +13,10 @@ public class BasicInfoResponse {
 	private final String studentNumber;
 	private final String major;
 	private final int totalCredit;
-	private final int takenCredit;
+	private final double takenCredit;
 
 	@Builder
-	private BasicInfoResponse(String name, String studentNumber, String major, int totalCredit, int takenCredit) {
+	private BasicInfoResponse(String name, String studentNumber, String major, int totalCredit, double takenCredit) {
 		this.name = name;
 		this.studentNumber = studentNumber;
 		this.major = major;
